Add bot sharding
