+++ .babelrc
@@ -3,5 +3,12 @@
     [
       "@babel/preset-env"
     ]
+  ],
+  "plugins": [
+    ["@babel/plugin-transform-runtime",
+      {
+        "regenerator": true
+      }
+    ]
   ]
 }
